+++ src/main/java/pl/tw/ksiegowosc/dto/CreateInvoiceResponse.java
@@ -1,0 +1,7 @@
+package pl.tw.ksiegowosc.dto;
+
+public record CreateInvoiceResponse(
+        String invoiceId,
+        String customerId
+) {
+}