+++ src/main/java/pl/tw/ksiegowosc/dto/MeritCreateInvoiceCustomer.java
@@ -1,0 +1,8 @@
+package pl.tw.ksiegowosc.dto;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+
+public record MeritCreateInvoiceCustomer(
+        @JsonProperty("Id") String id
+) {
+}